rename repository to gitrepo
