+++ src/gitrepo/manager.go
@@ -1,4 +1,4 @@
-package repository
+package gitrepo
 
 import (
 	"github.com/go-git/go-git/v5"
